Add analytics to post side recommendations
